+++ .gitignore
@@ -1,4 +1,5 @@
 .idea
+.vscode
 venv
 dist
 build
