+++ easy_manage/controllers/exceptions.py
@@ -1,3 +1,5 @@
+# pylint: disable=missing-docstring
+
 class NotInitializedError(Exception):
     pass
